FCOM-144: Some tweaks to file names and try-catch removal
