+++ Frends.Community.PgpDecrypt.sln
@@ -1,17 +1,17 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.27130.2020
+VisualStudioVersion = 15.0.26730.16
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SolutionItems", "SolutionItems", "{2238146A-EABB-4391-A460-0F2A786A31A5}"
 	ProjectSection(SolutionItems) = preProject
-		nuspec\Frends.Community.PgpDecrypt.nuspec = nuspec\Frends.Community.PgpDecrypt.nuspec
+		nuspec\Frends.Community.PgpDecryptFile.nuspec = nuspec\Frends.Community.PgpDecryptFile.nuspec
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.PgpDecrypt", "Frends.Community.PgpDecryptFile\Frends.Community.PgpDecrypt.csproj", "{97D07B5F-7D31-4D07-B89A-F4087EBEB82B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.PgpDecryptFile", "Frends.Community.PgpDecryptFile\Frends.Community.PgpDecryptFile.csproj", "{97D07B5F-7D31-4D07-B89A-F4087EBEB82B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.PgpDecrypt.Tests", "Frends.Community.PgpDecryptFile.Tests\Frends.Community.PgpDecrypt.Tests.csproj", "{067456C0-086C-46A8-B37F-1405717B7BFC}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.PgpDecryptFile.Tests", "Frends.Community.PgpDecryptFile.Tests\Frends.Community.PgpDecryptFile.Tests.csproj", "{067456C0-086C-46A8-B37F-1405717B7BFC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
