+++ Frends.Community.PgpDecryptFile/Frends.Community.PgpDecryptFile.csproj
@@ -54,7 +54,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="PgpDecrypt.cs" />
+    <Compile Include="PgpDecryptFile.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
