+++ nuspec/Frends.Community.PgpDecryptFile.nuspec
